+++ editor/editor-wikimarkup-transformer/src/parser/utils/attrs.ts
@@ -1,4 +1,6 @@
-import { parse as parseQuery } from 'querystring';
+// querystring-es3 is used because the browser implementation of querystring behaves differently than the NodeJS module
+// https://product-fabric.atlassian.net/browse/CS-2801
+import { parse as parseQuery } from 'querystring-es3';
 
 export function parseAttrs(
   str: string,