+++ media/media-table/src/util/generateRowValues.ts
@@ -1,26 +1,47 @@
-import { RowType } from '@atlaskit/dynamic-table/types';
+import {
+  RowType,
+  HeadCellType,
+  RowCellType,
+} from '@atlaskit/dynamic-table/types';
 import memoizeOne from 'memoize-one';
 
-function generateEmptyRows(length: number) {
+function generateEmptyRows(
+  emptyCells: RowCellType[],
+  length: number,
+): RowType[] {
   const itemsArray = Array.from({ length });
   return itemsArray.map(() => ({
-    cells: [],
+    cells: emptyCells,
   }));
 }
 
+export const generateEmptyRow = memoizeOne(
+  (headerCells: HeadCellType[]): RowCellType[] => {
+    return headerCells.map(cell => ({
+      key: cell.key,
+      content: '',
+    }));
+  },
+);
+
 export const prependRows = memoizeOne(
-  (itemsPerPage?: number, pageNumber?: number): RowType[] => {
+  (
+    emptyCells: RowCellType[],
+    itemsPerPage?: number,
+    pageNumber?: number,
+  ): RowType[] => {
     if (!itemsPerPage || !pageNumber) {
       return [];
     }
 
     const itemsCount = (pageNumber - 1) * itemsPerPage;
-    return generateEmptyRows(itemsCount);
+    return generateEmptyRows(emptyCells, itemsCount);
   },
 );
 
 export const appendRows = memoizeOne(
   (
+    emptyCells: RowCellType[],
     rowsLength: number,
     itemsPerPage?: number,
     pageNumber?: number,
@@ -35,19 +56,34 @@ export const appendRows = memoizeOne(
       (Math.ceil(totalItems / itemsPerPage) - pageNumber) * itemsPerPage -
       rowsDiff;
 
-    return generateEmptyRows(itemsCount);
+    return generateEmptyRows(emptyCells, itemsCount);
   },
 );
 
-export default function generateRowValues(
-  rowValues: RowType[],
-  itemsPerPage: number,
-  pageNumber: number,
-  totalItems: number,
-): RowType[] {
+export default function generateRowValues({
+  itemsPerPage,
+  pageNumber,
+  totalItems,
+  rowValues,
+  headerCells,
+}: {
+  itemsPerPage?: number;
+  pageNumber?: number;
+  totalItems?: number;
+  rowValues: RowType[];
+  headerCells: HeadCellType[];
+}): RowType[] {
+  const emptyCells = generateEmptyRow(headerCells);
+
   return [
-    ...prependRows(itemsPerPage, pageNumber),
+    ...prependRows(emptyCells, itemsPerPage, pageNumber),
     ...rowValues,
-    ...appendRows(rowValues.length, itemsPerPage, pageNumber, totalItems),
+    ...appendRows(
+      emptyCells,
+      rowValues.length,
+      itemsPerPage,
+      pageNumber,
+      totalItems,
+    ),
   ];
 }