+++ design-system/avatar/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/avatar
 
+## 18.0.1
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 18.0.0
 
 ### Major Changes