+++ design-system/blanket/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/blanket
 
+## 10.0.19
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 10.0.18
 
 ### Patch Changes