+++ design-system/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/breadcrumbs
 
+## 9.2.11
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 9.2.10
 
 ### Patch Changes