+++ design-system/code/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/code
 
+## 11.1.6
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 11.1.5
 
 ### Patch Changes