+++ design-system/empty-state/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/empty-state
 
+## 6.0.11
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 6.0.10
 
 ### Patch Changes