+++ design-system/flag/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/flag
 
+## 12.4.4
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 12.4.3
 
 ### Patch Changes