+++ design-system/form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/form
 
+## 7.3.1
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 7.3.0
 
 ### Minor Changes