+++ design-system/inline-dialog/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/inline-dialog
 
+## 12.1.14
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 12.1.13
 
 ### Patch Changes