+++ design-system/inline-edit/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/inline-edit
 
+## 10.0.33
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 10.0.32
 
 ### Patch Changes