+++ design-system/item/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/item
 
+## 11.0.3
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 11.0.2
 
 ### Patch Changes