+++ design-system/layer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/layer
 
+## 8.0.3
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 8.0.2
 
 ### Patch Changes