+++ design-system/motion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/motion
 
+## 0.2.5
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 0.2.4
 
 ### Patch Changes