+++ design-system/nps/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/nps
 
+## 7.0.2
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 7.0.1
 
 ### Patch Changes