+++ design-system/page-layout/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page-layout
 
+## 0.4.2
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 0.4.1
 
 ### Patch Changes