+++ design-system/page/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/page
 
+## 11.0.14
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 11.0.13
 
 ### Patch Changes