+++ design-system/pagination/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/pagination
 
+## 12.0.21
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 12.0.20
 
 ### Patch Changes