+++ design-system/portal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/portal
 
+## 3.1.9
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 3.1.8
 
 ### Patch Changes