+++ design-system/progress-indicator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/progress-indicator
 
+## 7.0.15
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 7.0.14
 
 ### Patch Changes