+++ design-system/rating/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/rating
 
+## 0.0.6
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 0.0.5
 
 ### Patch Changes