+++ design-system/section-message/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/section-message
 
+## 4.1.10
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 4.1.9
 
 ### Patch Changes