+++ design-system/tabs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tabs
 
+## 11.2.11
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 11.2.10
 
 ### Patch Changes