+++ design-system/tag-group/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/tag-group
 
+## 9.0.8
+
+### Patch Changes
+
+- [`54a9514fcf`](https://bitbucket.org/atlassian/atlassian-frontend/commits/54a9514fcf) - Build and supporting files will no longer be published to npm
+
 ## 9.0.7
 
 ### Patch Changes